Rename v6.0.0 to v6 upgrade
